add role + several fix error

+++ resources/views/review.blade.php
@@ -19,22 +19,24 @@
             </div>
             @endif
 
-            <form action="/review/proses" method="POST" enctype="multipart/form-data" style="padding-left: 5%">
-                {{ csrf_field() }}
+            @if (Auth::user()->role != 'admin')
+                <form action="/review/proses" method="POST" enctype="multipart/form-data" style="padding-left: 5%">
+                    {{ csrf_field() }}
 
-                <div class="form-group">
-                    <b>File Gambar</b><br/>
-                    <input type="file" name="file">
-                </div>
+                    <div class="form-group">
+                        <b>File Gambar</b><br/>
+                        <input type="file" name="file">
+                    </div>
 
-                <br>
-                <div class="form-group" style="padding-bottom: 3%">
-                    <b>Description</b> <br>
-                    <textarea class="form-control" name="description" cols="30" style="padding: 0.5% 2%;"></textarea>
-                </div>
+                    <br>
+                    <div class="form-group" style="padding-bottom: 3%">
+                        <b>Description</b> <br>
+                        <textarea class="form-control" name="description" cols="30" style="padding: 0.5% 2%;"></textarea>
+                    </div>
 
-                <input type="submit" value="Upload" class="btn btn-primary" style="background-color:green; color: white; padding: 0.5% 1%; margin-left:44%; margin-bottom: 2%">
-            </form>
+                    <input type="submit" value="Upload" class="btn btn-primary" style="background-color:green; color: white; padding: 0.5% 1%; margin-left:44%; margin-bottom: 2%">
+                </form>
+            @endif
 
             <h4 class="my-5" style="text-align: center;">Data</h4>
 
@@ -43,7 +45,9 @@
                     <tr>
                         <th width="1%">File</th>
                         <th>Description</th>
-                        <th width="1%">OPSI</th>
+                        @if (Auth::user()->role == 'admin')
+                            <th width="1%">OPSI</th>
+                        @endif
                     </tr>
                 </thead>
                 <tbody>
@@ -51,7 +55,9 @@
                     <tr>
                         <td><img width="400vw" src="{{ url('/images/'.$g->file) }}"></td>
                         <td style="text-align: center; width: 1rem">{{$g->description}}</td>
-                        <td><a class="btn btn-danger" href="review/delete{{ $g->id }}">HAPUS</a></td>
+                        @if (Auth::user()->role == 'admin')
+                            <td><a class="btn btn-danger" href="review/delete{{ $g->id }}">HAPUS</a></td>
+                        @endif
                     </tr>
 
                     @empty
